Fix admin role forgery for user profiles
I mistakenly thought that I had to replace the block "role=user       " with "role=admin      " when in reality I could let the "role=" portion remain in the previous block and just replace user with admin. That made me make a weird deserializer that allowed = to exist in user input and not be escaped at all which is clearly not correct.

I would have liked to use url.EscapeQuery instead of replacing "&" and "=" with " ", but this would cause my padding bytes to be escaped as well and I wouldn't be able to perform this attack.